// src/sdk/generated/nfl.d.ts
export interface paths {
// ...
    "/v3/nfl/scores/{format}/Stadiums": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Stadiums
         * @description Returns all stadiums in the NFL with capacity, surface, latitude/longitude, city and state (and, where applicable, country.)
         */
        get: operations["nfl_v3_scores_stadiums"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/nfl/scores/{format}/ScoresByDate/{date}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Games - by Date [Live & Final]
         * @description Full scores and gameday info delivered live and post-game. Live data includes down and distance, as well as game clock. Gameday info includes referee, weather, TV channel etc.
         */
        get: operations["nfl_v3_scores_games___by_date__live___final"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/nfl/projections/{format}/PlayerGameProjectionStatsByWeek/{season}/{week}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Projected Player Game Stats - by Week
         * @description SportsDataIO's proprietary projections, including DFS salary information and injuries, for fantasy players, called by week.
         */
        get: operations["nfl_v3_projections_projected_player_game_stats___by_week"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/nfl/odds/{format}/PreGameOddsLineMovement/{scoreid}/{sportsbookgroup}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Pre-Game and Period Game Odds Line Movement [Sportsbook Group]
         * @description Returns the game odds (for games in a given week and season. In this endpoint both full-game and partial-game odds are included. Returns the full line movement for the given game. This endpoint has a longer cache as it is meant for historical data/line movement rather than the most up to the second line. A Sportsbook Group must be specified as a URL parameter.
         */
        get: operations["nfl_v3_odds_pre_game_and_period_game_odds_line_movement__sportsbook_group"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/nfl/odds/{format}/InGameLineMovement/{scoreid}/{sportsbookgroup}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * In-Game Odds Line Movement [Sportsbook Group]
         * @description Returns in-play game odds (spread, moneyline, total) for games in a given week and season. Returns the full line movement for the given game. This endpoint has a longer cache as it is meant for historical data/line movement rather than the most up to the second line. A Sportsbook Group must be specified as a URL parameter.
         */
        get: operations["nfl_v3_odds_in_game_odds_line_movement__sportsbook_group"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/nfl/odds/{format}/BettingMarketsByGameID/{gameid}/{sportsbookgroup}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Betting Markets - by Game [Sportsbook Group]
         * @description Returns the markets of all available types (e.g. Player Props, Team Props) and available outcomes for a given GameID. A Sportsbook Group must be specified as a URL parameter.
         */
        get: operations["nfl_v3_odds_betting_markets___by_game__sportsbook_group"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/nfl/stats/{format}/BoxScoreByScoreIDV3/{scoreid}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Box Score [Live & Final]
         * @description Full statistical information for a specified game, down to the team and player stat level, delivered live during the game, called per individual game.
         */
        get: operations["nfl_v3_stats_box_score__live___final"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/nfl/pbp/{format}/PlayByPlay/{gameid}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Play By Play [Live & Final]
         * @description Each invididual play, its type and outcome, complete with player and team stats down to the play level, delivered live in real-time, called by game.
         */
        get: operations["nfl_v3_pbp_play_by_play__live___final"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}